This patch fix man miscellenous compilation warnings. Signed-off-by: CAI Qian <caiqian@cclom.cn>.

+++ testcases/kernel/fs/openfile/openfile.c
@@ -33,6 +33,7 @@
 #include <pthread.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include <unistd.h>
 
 
@@ -136,7 +137,8 @@ int main(int argc, char *argv[])
 
 	/* Create threads */
 	for (i=0; i<numthreads; i++)
-		if (pthread_create(&th_id, (pthread_attr_t *)NULL, threads, (void *)i)) {
+		if (pthread_create(&th_id, (pthread_attr_t *)NULL, threads,
+				(void *)(uintptr_t)i)) {
 			perror("FAIL - failed creating a pthread; increase limits");
 			fclose(fd);
 			unlink(filename);
@@ -184,7 +186,7 @@ int main(int argc, char *argv[])
 /* threads: Each thread opens the files specified */
 void * threads(void* thread_id_)
 {
-  int thread_id=(int)thread_id_;
+  int thread_id=(uintptr_t)thread_id_;
     	char  errmsg[80];
         FILE  *fd;
 	int   i;
